+++ xalpha/__init__.py
@@ -10,4 +10,5 @@ from xalpha.multiple import mul, mulfix, imul
 from xalpha.realtime import rfundinfo, review
 from xalpha.record import record, irecord
 from xalpha.trade import trade, itrade
-from xalpha.universal import get_daily, get_rt
+from xalpha.universal import get_daily, get_rt, set_backend
+from xalpha.provider import show_providers