+++ xalpha/universal.py
@@ -5,14 +5,34 @@ for almost everything in the market
 """
 
 import os
+import sys
 import datetime as dt
+import numpy as np
 import pandas as pd
 from bs4 import BeautifulSoup
 from functools import wraps
 from sqlalchemy import exc
+from dateutil.relativedelta import relativedelta
+
+try:
+    from jqdatasdk import (
+        get_index_weights,
+        query,
+        get_fundamentals,
+        valuation,
+        get_query_count,
+    )
+
+except ImportError:
+    pass
 
 from xalpha.info import fundinfo, mfundinfo
-from xalpha.cons import rget, rpost
+from xalpha.cons import rget, rpost, rget_json, rpost_json
+from xalpha.provider import data_source
+
+
+thismodule = sys.modules[__name__]
+xamodule = sys.modules["xalpha"]
 
 
 def today_obj():
@@ -142,6 +162,19 @@ def get_investing_id(suburl):
     return pid
 
 
+def _variate_ua():
+    last = 20 + np.random.randint(20)
+    ua = []
+    ua.append(
+        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/537.36 (KHTML, like Gecko)"
+    )
+    ua.append(
+        "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1"
+    )
+    choice = np.random.randint(2)
+    return ua[choice][:last]
+
+
 def get_rmb(start=None, end=None, prev=360, currency="USD/CNY"):
     """
     获取人民币汇率中间价
@@ -165,38 +198,50 @@ def get_rmb(start=None, end=None, prev=360, currency="USD/CNY"):
     end_str = end_obj.strftime("%Y-%m-%d")
     count = (end_obj - start_obj).days + 1
     rl = []
+    # API 很奇怪，需要经常变 UA 才好用
+
+    headers = {
+        "Referer": "http://www.chinamoney.com.cn/chinese/bkccpr/",
+        "Origin": "http://www.chinamoney.com.cn",
+        "Host": "www.chinamoney.com.cn",
+        "X-Requested-With": "XMLHttpRequest",
+    }
+
     if count <= 360:
-        r = rpost(
+        headers.update({"user-agent": _variate_ua()})
+        r = rpost_json(
             url.format(start_str=start_str, end_str=end_str, currency=currency),
-            headers={"user-agent": "Mozilla/5.0"},
+            headers=headers,
         )
-        rl.extend(r.json()["records"])
+        rl.extend(r["records"])
     else:  # data more than 1 year cannot be fetched once due to API limitation
         sepo_obj = end_obj
         sepn_obj = sepo_obj - dt.timedelta(360)
         #         sep0_obj = end_obj - dt.timedelta(361)
         while sepn_obj > start_obj:  # [sepn sepo]
-            r = rpost(
+            headers.update({"user-agent": _variate_ua()})
+            r = rpost_json(
                 url.format(
                     start_str=sepn_obj.strftime("%Y-%m-%d"),
                     end_str=sepo_obj.strftime("%Y-%m-%d"),
                     currency=currency,
                 ),
-                headers={"user-agent": "Mozilla/5.0"},
+                headers=headers,
             )
-            rl.extend(r.json()["records"])
+            rl.extend(r["records"])
 
             sepo_obj = sepn_obj - dt.timedelta(1)
             sepn_obj = sepo_obj - dt.timedelta(360)
-        r = rpost(
+        headers.update({"user-agent": _variate_ua()})
+        r = rpost_json(
             url.format(
                 start_str=start_obj.strftime("%Y-%m-%d"),
                 end_str=sepo_obj.strftime("%Y-%m-%d"),
                 currency=currency,
             ),
-            headers={"user-agent": "Mozilla/5.0"},
+            headers=headers,
         )
-        rl.extend(r.json()["records"])
+        rl.extend(r["records"])
     data = {"date": [], "close": []}
     for d in rl:
         data["date"].append(pd.Timestamp(d["date"]))
@@ -216,7 +261,7 @@ def get_fund(code):
     return df[["date", "close"]]
 
 
-def get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
+def _get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
     """
     universal fetcher for daily historical data of literally everything has a value in market.
     数据来源包括天天基金，雪球，英为财情，外汇局官网
@@ -236,6 +281,11 @@ def get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
             6. 对于国内发行的基金，使用基金代码，同时开头添加 F。
 
             7. 对于国内发行的货币基金，使用基金代码，同时开头添加 M。（全部按照净值数据处理）
+
+            8. 形如 peb-000807.XSHG 格式的数据，可以返回每周的指数估值情况，需要 enable 聚宽数据源方可查看。
+
+            9. 形如 iw-000807.XSHG 格式的数据，可以返回每月的指数成分股和实时权重，需要 enable 聚宽数据源方可查看。
+
     :param start: str. "20200101", "2020/01/01", "2020-01-01" are all legal. The starting date of daily data.
     :param end: str. format is the same as start. The ending date of daily data.
     :param prev: Optional[int], default 365. If start is not specified, start = end-prev.
@@ -249,15 +299,16 @@ def get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
     else:
         end_obj = dstr2dobj(end)
     if not start:
-        start_obj = end_obj - dt.timedelta(prev)
+        start_obj = end_obj - dt.timedelta(days=prev)
     else:
         start_obj = dstr2dobj(start)
+
     if not _from:
         if code.startswith("SH") or code.startswith("SZ"):
             _from = "xueqiu"
         elif code.endswith("/CNY") or code.startswith("CNY/"):
             _from = "zjj"
-        elif len(code.split("/")) > 1:
+        elif len(code[1:].split("/")) == 2:
             _from = "cninvesting"
             code = get_investing_id(code)
         elif code.isdigit():
@@ -267,6 +318,9 @@ def get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
         elif code.startswith("HK") and code[2:].isdigit() and len(code) == 7:
             _from = "xueqiu"
             code = code[2:]
+        elif len(code.split("-")) == 2 and len(code.split("-")[0]) <= 3:
+            # peb-000807.XSHG
+            _from = code.split("-")[0]
         else:
             _from = "xueqiu"
 
@@ -289,6 +343,20 @@ def get_daily(code, start=None, end=None, prev=365, _from=None, **kws):
         df = get_fund(code)
         df = df[df.date <= end_str]
         df = df[df.date >= start_str]
+        return df
+    elif _from == "peb":
+        df = _get_peb_range(code=code.split("-")[1], start=start_str, end=end_str)
+        if len(df) > 0:
+            df = df[df.date <= end_str]
+            df = df[df.date >= start_str]
+        return df
+    elif _from == "iw":
+        df = _get_index_weight_range(
+            code=code.split("-")[1], start=start_str, end=end_str
+        )
+        if len(df) > 0:
+            df = df[df.date <= end_str]
+            df = df[df.date >= start_str]
         return df
 
 
@@ -396,7 +464,7 @@ _cached_data = {}
 
 def reset_cache():
     """
-    clear all cache of daily data
+    clear all cache of daily data, used for :func:`cached`, deprecated.
 
     :return: None.
     """
@@ -406,6 +474,8 @@ def reset_cache():
 
 def cached(s):
     """
+    **Deprecated**, use :func:`cachedio` instead, where ``backend="memory"``.
+
     Usage as follows:
 
     .. code-block:: python
@@ -465,9 +535,12 @@ def cached(s):
 
 def cachedio(**ioconf):
     """
-    用法类似:func:`cached`，只不过是硬盘级别的缓存
+    用法类似:func:`cached`，通用透明缓存器，用来作为 (code, start, end ...) -> pd.DataFrame 形式函数的缓存层，
+    避免重复爬取已有数据。
 
-    :param **ioconf: 可选关键字参数 backend: csv or sql, path: csv 文件夹或 sql engine， refresh True 会刷新结果，重新爬取， prefix 是 key 前统一部分
+    :param **ioconf: 可选关键字参数 backend: csv or sql or memory,
+        path: csv 文件夹或 sql engine， refresh True 会刷新结果，重新爬取, default False，
+        prefix 是 key 前统一部分, 缓存 hash 标志
     :return:
     """
 
@@ -480,22 +553,29 @@ def cachedio(**ioconf):
                 code = kws.get("code")
             date = ioconf.get("date", "date")
             key = kws.get("key", code)
+            key = key.replace("/", " ")
             start = kws.get("start", None)
             end = kws.get("end", None)
+            prev = kws.get("prev", None)
             prefix = ioconf.get("prefix", "")
             key = prefix + key
+            # print("xdebug: %s" % ioconf.get("backend", "no"))
+            if not prev:
+                prev = 365
             if not end:
                 end_obj = today_obj()
             else:
                 end_obj = dt.datetime.strptime(
                     end.replace("/", "").replace("-", ""), "%Y%m%d"
                 )
+
             if not start:
-                start_obj = end_obj - dt.timedelta(days=365)
+                start_obj = end_obj - dt.timedelta(days=prev)
             else:
                 start_obj = dt.datetime.strptime(
                     start.replace("/", "").replace("-", ""), "%Y%m%d"
                 )
+
             start_str = start_obj.strftime("%Y%m%d")
             end_str = end_obj.strftime("%Y%m%d")
             backend = ioconf.get("backend")
@@ -506,6 +586,9 @@ def cachedio(**ioconf):
             else:
                 if backend == "csv":
                     key = key + ".csv"
+                elif backend == "memory":
+                    if not getattr(thismodule, "cached_dict", None):
+                        setattr(thismodule, "cached_dict", {})
                 if refresh:
                     df0 = f(*args, **kws)
 
@@ -515,6 +598,8 @@ def cachedio(**ioconf):
                             df0 = pd.read_csv(os.path.join(path, key))
                         elif backend == "sql":
                             df0 = pd.read_sql(key, path)
+                        elif backend == "memory":
+                            df0 = getattr(thismodule, "cached_dict")[key]
                         else:
                             raise ValueError("no %s option for backend" % backend)
                         df0[date] = pd.to_datetime(df0[date])
@@ -537,20 +622,295 @@ def cachedio(**ioconf):
                             if len(df2) > 0:
                                 df0 = df0.append(df2, ignore_index=True)
 
-                    except (FileNotFoundError, exc.ProgrammingError):
+                    except (FileNotFoundError, exc.ProgrammingError, KeyError):
                         df0 = f(*args, **kws)
 
-                if len(df0) > 0:
+                if df0 is not None and len(df0) > 0:
                     if backend == "csv":
                         df0.to_csv(os.path.join(path, key), index=False)
                     elif backend == "sql":
                         df0.to_sql(key, con=path, if_exists="replace", index=False)
+                    elif backend == "memory":
+                        d = getattr(thismodule, "cached_dict")
+                        d[key] = df0
 
-            df0 = df0[df0["date"] <= end_str]
-            df0 = df0[df0["date"] >= start_str]
+            if df0 is not None:
+                df0 = df0[df0["date"] <= end_str]
+                df0 = df0[df0["date"] >= start_str]
 
             return df0
 
         return wrapper
 
     return cached
+
+
+@data_source("jq")
+def _get_index_weight_range(code, start, end):
+    start_obj = dt.datetime.strptime(start.replace("-", "").replace("/", ""), "%Y%m%d")
+    end_obj = dt.datetime.strptime(end.replace("-", "").replace("/", ""), "%Y%m%d")
+    start_m = start_obj.replace(day=1)
+    if start_m < start_obj:
+        start_m = start_m + relativedelta(months=1)
+    end_m = end_obj.replace(day=1)
+    if end_obj < end_m:
+        end_m = end_m - relativedelta(months=1)
+    d = start_m
+
+    df = pd.DataFrame({"code": [], "weight": [], "display_name": [], "date": []})
+    while True:
+        if d > end_m:
+
+            df["date"] = pd.to_datetime(df["date"])
+            return df
+        print("call", d, " ", code)
+        df0 = get_index_weights(index_id=code, date=d.strftime("%Y-%m-%d"))
+        df0["code"] = df0.index
+        df = df.append(df0, ignore_index=True)
+        d = d + relativedelta(months=1)
+
+
+@data_source("jq")
+def _get_peb_range(code, start, end):  # 盈利，净资产，总市值
+    """
+    获取指定指数一段时间内的 pe pb 值。
+
+    :param code: 聚宽形式指数代码。
+    :param start:
+    :param end:
+    :return: pd.DataFrame
+    """
+    data = {"date": [], "pe": [], "pb": []}
+    for d in pd.date_range(start=start, end=end, freq="W-FRI"):
+        data["date"].append(d)
+        print("compute pe pb on %s" % d)
+        r = get_peb(code, date=d.strftime("%Y-%m-%d"))
+        data["pe"].append(r["pe"])
+        data["pb"].append(r["pb"])
+    return pd.DataFrame(data)
+
+
+def set_backend(**ioconf):
+    """
+    设定 xalpha get_daily 函数的缓存后端，默认为内存。 ioconf 参数设置可参考 :func:`cachedio`
+
+    :param ioconf:
+    :return: None.
+    """
+
+    if not ioconf:
+        ioconf = {"backend": "memory"}
+    get_daily = cachedio(**ioconf)(_get_daily)
+    prefix = ioconf.get("prefix", "")
+    ioconf["prefix"] = "iw-" + prefix
+    get_index_weight_range = cachedio(**ioconf)(_get_index_weight_range)
+    ioconf["prefix"] = "peb-" + prefix
+    get_peb_range = cachedio(**ioconf)(_get_peb_range)
+    setattr(thismodule, "get_daily", get_daily)
+    setattr(xamodule, "get_daily", get_daily)
+    setattr(thismodule, "get_index_weight_range", get_index_weight_range)
+    setattr(thismodule, "get_peb_range", get_peb_range)
+    setattr(thismodule, "ioconf", ioconf)
+
+
+set_backend()
+
+
+@data_source("jq")
+def get_peb(index, date=None, table=False):
+    """
+    获取指数在指定日期的 pe 和 pb。采用当时各公司的最新财报和当时的指数成分股权重加权计算。
+
+    :param index: str. 聚宽形式的指数代码。
+    :param date: str. %Y-%m-%d
+    :param table: Optioanl[bool], default False. True 时返回整个计算的 DataFrame，用于 debug。
+    :return: Dict[str, float]. 包含 pe 和 pb 值的字典。
+    """
+    middle = dt.datetime.strptime(
+        date.replace("/", "").replace("-", ""), "%Y%m%d"
+    ).replace(day=1)
+    iwdf = get_index_weight_range(
+        index,
+        start=(middle - dt.timedelta(days=5)).strftime("%Y-%m-%d"),
+        end=(middle + dt.timedelta(days=5)).strftime("%Y-%m-%d"),
+    )
+    q = query(valuation).filter(valuation.code.in_(list(iwdf.code)))
+    #     df = get_fundamentals(q, date)
+    df = get_fundamentals(q, date=date)
+    df = df.merge(iwdf, on="code")
+    df["e"] = df["weight"] / df["pe_ratio"]
+    df["b"] = df["weight"] / df["pb_ratio"]
+    df["p"] = df["weight"]
+    tote = df.e.sum()
+    totb = df.b.sum()
+    if table:
+        return df
+    return {"pe": 100.0 / tote, "pb": 100.0 / totb}
+
+
+def _convert_code(code):
+    """
+    将聚宽形式的代码转化为 xalpha 形式
+
+    :param code:
+    :return:
+    """
+    no, mk = code.split(".")
+    if mk == "XSHG":
+        return "SH" + no
+    elif mk == "XSHE":
+        return "SZ" + no
+
+
+def _inverse_convert_code(code):
+    """
+    将 xalpha 形式的代码转化为聚宽形式
+
+    :param code:
+    :return:
+    """
+
+    if code.startswith("SH"):
+        return code[2:] + ".XSHG"
+    elif code.startswith("SZ"):
+        return code[2:] + ".XSHE"
+
+
+class PEBHistory:
+    """
+    对于指数历史 PE PB 的封装类
+    """
+
+    indexs = {
+        "000016.XSHG": ("上证50", "2012-01-01"),
+        "000300.XSHG": ("沪深300", "2012-01-01"),
+        "000905.XSHG": ("中证500", "2012-01-01"),
+        "000922.XSHG": ("中证红利", "2012-01-01"),
+        "399006.XSHE": ("创业板指", "2012-01-01"),
+        "000992.XSHG": ("全指金融", "2012-01-01"),
+        "000991.XSHG": ("全指医药", "2012-01-01"),
+        "399932.XSHE": ("中证消费", "2012-01-01"),
+        "000831.XSHG": ("500低波", "2013-01-01"),
+        "000827.XSHG": ("中证环保", "2013-01-01"),
+        "000978.XSHG": ("医药100", "2012-01-01"),
+        "399324.XSHE": ("深证红利", "2012-01-01"),
+        "399971.XSHE": ("中证传媒", "2014-07-01"),
+        "000807.XSHG": ("食品饮料", "2013-01-01"),
+        "000931.XSHG": ("中证可选", "2012-01-01"),
+        "399812.XSHE": ("养老产业", "2016-01-01"),
+    }
+
+    def __init__(self, code, start=None, end=None):
+        """
+
+        :param code: str. 形式可以是 399971.XSHE 或者 SH000931
+        :param start: Optional[str]. %Y-%m-%d, 估值历史计算的起始日。
+        :param end: Dont use, only for debug
+        """
+        yesterday_str = (dt.datetime.now() - dt.timedelta(days=1)).strftime("%Y-%m-%d")
+        if len(code.split(".")) == 2:
+            self.code = code
+            self.scode = _convert_code(code)
+        else:
+            self.scode = code
+            self.code = _inverse_convert_code(self.scode)
+        if self.code in self.indexs:
+            self.name = self.indexs[self.code][0]
+            if not start:
+                start = self.indexs[self.code][1]
+        else:
+            try:
+                self.name = get_rt(self.scode)["name"]
+            except:
+                self.name = self.scode
+            if not start:
+                start = "2012-01-01"  # 可能会出问题，对应指数还未有数据
+        self.start = start
+        if not end:
+            end = yesterday_str
+        self.df = get_peb_range(self.code, start=self.start, end=end)
+        self.ratio = None
+        self.pep = [
+            round(i, 3) for i in np.percentile(self.df.pe, np.arange(0, 110, 10))
+        ]
+        self.pbp = [
+            round(i, 3) for i in np.percentile(self.df.pb, np.arange(0, 110, 10))
+        ]
+
+    def percentile(self):
+        """
+        打印 PE PB 的历史十分位对应值
+
+        :return:
+        """
+        print("PE 历史分位:\n")
+        print(*zip(np.arange(0, 110, 10), self.pep), sep="\n")
+        print("\nPB 历史分位:\n")
+        print(*zip(np.arange(0, 110, 10), self.pbp), sep="\n")
+
+    def v(self, y="pe"):
+        """
+        pe 或 pb 历史可视化
+
+        :param y: Optional[str]. "pe" (defualt) or "pb"
+        :return:
+        """
+        return self.df.plot(x="date", y=y)
+
+    def fluctuation(self):
+        if not self.ratio:
+            d = self.df.iloc[-1]["date"]
+            oprice = get_daily(code=self.scode, end=d.strftime("%Y%m%d"), prev=20).iloc[
+                -1
+            ]["close"]
+            nprice = get_rt(self.scode)["current"]
+            self.ratio = nprice / oprice
+        return self.ratio
+
+    def current(self, y="pe"):
+        """
+        返回实时的 pe 或 pb 绝对值估计。
+
+        :param y: Optional[str]. "pe" (defualt) or "pb"
+        :return: float.
+        """
+        return round(self.df.iloc[-1][y] * self.fluctuation(), 3)
+
+    def current_percentile(self, y="pe"):
+        """
+        返回实时的 pe 或 pb 历史百分位估计
+
+        :param y: Optional[str]. "pe" (defualt) or "pb"
+        :return: float.
+        """
+        df = self.df
+        d = len(df)
+        u = len(df[df[y] < self.current(y)])
+        return round(u / d * 100, 2)
+
+    def summary(self):
+        """
+        打印现在估值的全部分析信息。
+
+        :return:
+        """
+        print("指数%s估值情况\n" % self.name)
+        if dt.datetime.strptime(self.start, "%Y-%m-%d") > dt.datetime(2015, 1, 1):
+            print("(历史数据较少，仅供参考)\n")
+        #         self.percentile()
+        print(
+            "现在 PE 绝对值 %s, 相对分位 %s%%，距离最低点 %s %%\n"
+            % (
+                self.current("pe"),
+                self.current_percentile("pe"),
+                round((self.current("pe") - self.pep[0]) / self.current("pe") * 100, 1),
+            )
+        )
+        print(
+            "现在 PB 绝对值 %s, 相对分位 %s%%，距离最低点 %s %%\n"
+            % (
+                self.current("pb"),
+                self.current_percentile("pb"),
+                round((self.current("pb") - self.pbp[0]) / self.current("pb") * 100, 1),
+            )
+        )